Step 6+7: package restructure and README overhaul
Step 6 — autovision/ package
- Move scraper, model, trainer, pipeline into autovision/ package
- Extract TrainConfig into autovision/config.py (single source of truth
  for all hyperparameters, importable without pulling in torch)
- Add autovision/__init__.py exposing run(), train(), predict(), TrainConfig
- Use relative imports (.config, .model, .trainer, .scraper) inside package
- Update cli.py and app.py to import from autovision.*
- Add [tool.setuptools.packages.find] to pyproject.toml
- Delete root-level model.py, scraper.py, trainer.py, pipeline.py

Step 7 — README rewrite
- New title, one-line pitch, Python/PyTorch/MIT badges
- Quick-start section (4 commands: install, train, predict, demo)
- ASCII pipeline diagram showing the full data flow
- Training details table (backbone, loss, optimizer, augmentation, device)
- Full CLI reference for all four commands with options tables
- Python API section showing both one-call (run()) and step-by-step usage
- Project layout map
